+++ APBuild/Library/hlslparser/GLSLGenerator.cpp
@@ -475,7 +475,7 @@ void GLSLGenerator::OutputExpression(HLSLExpression* expression, const HLSLType*
             memberAccess->object->expressionType.baseType == HLSLBaseType_Uint)
         {
             // Handle swizzling on scalar values.
-            int swizzleLength = strlen(memberAccess->field);
+            int swizzleLength = (int)strlen(memberAccess->field);
             if (swizzleLength == 2)
             {
                 m_writer.Write("%s", m_scalarSwizzle2Function);
